// src/pages/organize/risk/search/data.d.ts
export interface TableListItem {
  key: number;
  disabled?: boolean;
  href: string;
  avatar: string;
  name: string;
  title: string;
  owner: string;
  desc: string;
  callNo: number;
  status: number;
  updatedAt: Date;
  createdAt: Date;
  progress: number;
}

export interface TableListPagination {
  total: number;
  pageSize: number;
  current: number;
}

export interface TableListData {
  list: TableListItem[];
  pagination: Partial<TableListPagination>;
}

export interface TableListParams {
  sorter: string;
  status: string;
  name: string;
  pageSize: number;
  currentPage: number;
}

export interface QiyeParams {
  key: number;
  name: string;
  no: string;
  type: string;
  sj: string;
}

export interface GerenParams {
  key: number;
  name: string;
  idcard: string;
  phone: string;
  sex: number;
  sj: string;
}